feat: dedicated helper to create a font object form OUDS token (#1550)
Closes #1550

Suggested-by: François Suc <francois.suc@orange.com>
