+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Helper to forge a font object with OUDS typography tokens (Orange-OpenSource/ouds-ios#1550)
 - Flag to limit the height of the `text area` component (Orange-OpenSource/ouds-ios#1528)
 - `View modifier` for fonts using only token (Orange-OpenSource/ouds-ios#1534)
